feat: start doing loaders
