--- src/routes/graphql/types/interfaces.ts
@@ -1,14 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-
-import DataLoader from 'dataloader';
-
-type IDataLoader = DataLoader<string, unknown>;
-
-export interface Loaders {
-  user: IDataLoader;
-}
-
-export interface Context {
-  db: PrismaClient;
-  loaders: Loaders;
-}
